Moved interactive traceApertures() from geminidr.gmos to geminidr.core
